requirements.txt
azure-functions
pyodbc
msrestazure
azure-identity
azure-keyvault-secrets